- extracted encoding fix into its own concern
